change location of target scripts
We need a better place to put our target scripts when we start running
on read-only images. This moves them to /home/phablet/bin. For
classic image testing on the target, we'll still need the in $PATH, so
jenkins.sh has a hack to copy them over.

Also fix a small issue when run locally this causes
a problem because of the "bzr export utils ...". This renames it to
something safe.