introduce concept of a mega job
This puts all the tests into a single giant job that's assigned
to a slave-node based on a device and not a server that then
requires locks-and-latches.

This was implemented in C-style type IFDEF approach. The reason
it was done this way is because MEGA=1 will ultimately become
the only way we run jobs. So this will make it easier to know
what code we can delete in the future.