fix bugs found in run-autopilot.sh
1: When unlock screen failed the whole script was ended. This makes sure
we just fail that single app test and continue trying other apps.

2: The reboot_wait logic isn't resiliant enough. Before we had "mega"
jobs (ie jenkins.sh) it was okay if a test occassionally failed to
get active networking. However, in this new script that failure can
cause all the remaining tests to not get run. The idea is to create
a reboot-and-wait script that will retry 3 times before giving up.

3: Update the system settle "classname" for junit. The issue I'm
seeing is that the junit results viewer in jenkins isn't correlating
the settle failures with the testsuite it happened in. This makes sure
the 2 are tied together.