add a "device" command to help manage device pool reservations
This adds a new command that provides a lock-safe mechanism to
give exclusive access to a device.

= How it Works
There's a "device pool" directory. Under this directory callers will
obtain mutually exclusive access via an flock to a file in this
directory. Once locked, the code can scan the directory. Each file
under this directory represents a device and includes the device's
ANDROID_SERIAL. A device is toggled between locked/free by its name.
eg: mako-01 will be the name when its free. When its in use it will
be named mako-01.lck. No changes are made to this directory unless
the caller has the flock.

= How to Use It
1) ./device populate -c <your jenkins config file>
This will create a device pool with every device defined in your
config file.

2) aquire a lock to a device
There are 3 ways:
  a) by name: ./device-get --name mako-01

  b) by type: ./device-get --type mako

  c) by JOBNAME environment variable:
     JOBNAME=saucy-touch_ro-mako-smoke-default ./device get

A token file will be created named .device-<yourpid>.  This is
used by step 3 below.

3) release your lock on a device:
  ./device free