fix issue in unlock_screen.sh and powerd-cli
We've been shifting around where powerd-cli is called. revno 77
moved the call into unlock_screen.sh. The problem is that the adb-shell
on the target won't exit until the two powerd-cli pids are killed. This
kills them once the screen is unlocked. Based on my testing this seems
to be adequate for ensuring the screen is on while the autopilot tests
are executed.