change order of operations in provision.sh
This re-orders things to make provision.sh a tiny bit faster and
also potentially more reliable. The thinking was:

we were waiting a bit because we were trying to set up networking.
phablet-click-test-setup does a lot of work on the host side first,
so if we call that earlier we don't have to sleep as long. We then
make the CUSTOMIZE piece last since it causes a reboot.