add some logging with timestamps
We have statsd metrics in run-smoke for provisioning as a whole,
but I saw an issue today where we need a little more debug information
to see where a provisioning was taking too long. This might be better
as a statsd metric when we convert this script to python. However, this
is a quick fix to get us out of trouble for now.