update jenkins job
This creates a properly named job with an explanation of the parameters.
We run everything via UTAH now instead of using the autopilot runner logic
so that the dashboard can parse all the results.

In the future when the dashboard can deal with non-utah (or we update
our jobs to convert junit|subunit to utah) then we can start enabling
that parameter.