move autopilot test installation out of provisioning
Currently our touch image testing logic is to install all tests
required after provisioning and before we start running the tests.
We need to instead install the packages required by a testsuite
just before running the test, and then uninstall the packages after.

The rationale behind this is because ofono-simd is getting installed
but is affecting other tests.

Since the test now has to include setup/teardown logic 2 new results
are being added to the results.xml file