make concurrency and parallel jobs easier to manage
We currently run daily image on its own and then request all
the qa jobs to run at the same time. This works okay in Jenkins,
but we'll soon need to be able to distribute the daily image
testing across a set of slaves (for emulator mostly but should
help everything).

This provides the basic ground work by switching us over to using
the build-flow-plugin in Jenkins:
 https://wiki.jenkins-ci.org/display/JENKINS/Build+Flow+Plugin

This MP currently changes nothing in how our testing runs, but will
make it much easier for me to do my follow on MP's for emulator support.