dashboard api: remove race condition
The dashboard itself is changing to prevent the creation of duplicate
SmokeImage objects. We'll now get this as an http exception when trying
to create an object. We need to handle that by then doing a GET call
to find the SmokeImage created by the other thread that won the race.