add emulator support to touch scripts
This allows you to test an emulator by hand using a command like:

 IMAGE_OPT="--channel devel-proposed" USE_EMULATOR=1 ./scripts/run-smoke -a friends_app

= Notable changes:

run-autopilot-tests.sh: this printed out the image info at the start
of a test assuming the device is available. This doesn't work with the
emulator, so I moved this to run-smoke where we first know about the
image-info.

run-smoke - we cant detect the device since it hasn't been created so
we have to fake this out

The remaining changes were basically setting up no-op stuff for things
the emulator doesn't need to deal with.